Only display debug info if LURCH_DEBUG is set
If LURCH_DEBUG is not set or is set to 0, llua_panic and die will no
longer emit stack traces; but will instead abort with a message
informing the user that they must set LURCH_DEBUG for more information.

llua_panic simply check to see if LURCH_DEBUG is set before printing
backtraces, while die() uses the value of LURCH_DEBUG as the number of
stack frames to print.