Change format of escape sequences; fix writeline
Previously, the sequence for changing the foreground/background colors
was '\x1b2<color>', where <color> was the color in binary form; however,
this caused issues where <color> was 10 (which translated to a newline)
or some other control/nonprinting character...

This commit changes the format ot '\x1b2XXXm', where XXX is the color in
decimal form.

Also, writeline() was fixed to move the column cursor forward by the
exact display width of the character that was printed, which is
calculated using wcwidth().